Organize as solution, add test project
