+++ TenantFile/Api/Program.cs
@@ -7,7 +7,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace TenantFileApi
+namespace TenantFile.Api
 {
     public class Program
     {
