+++ TenantFile/Api/Startup.cs
@@ -12,7 +12,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace TenantFileApi
+namespace TenantFile.Api
 {
     public class Startup
     {
