+++ TenantFile/TenantFile.sln
@@ -1,0 +1,48 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TenantFile.Api", "Api\TenantFile.Api.csproj", "{CE371E98-CF0F-4BF2-8229-C3FD8DDC794C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api.UnitTests", "Api.UnitTests\Api.UnitTests.csproj", "{C947E269-7D43-49ED-99C5-DB0C084B217E}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CE371E98-CF0F-4BF2-8229-C3FD8DDC794C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CE371E98-CF0F-4BF2-8229-C3FD8DDC794C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CE371E98-CF0F-4BF2-8229-C3FD8DDC794C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{CE371E98-CF0F-4BF2-8229-C3FD8DDC794C}.Debug|x64.Build.0 = Debug|Any CPU
+		{CE371E98-CF0F-4BF2-8229-C3FD8DDC794C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{CE371E98-CF0F-4BF2-8229-C3FD8DDC794C}.Debug|x86.Build.0 = Debug|Any CPU
+		{CE371E98-CF0F-4BF2-8229-C3FD8DDC794C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CE371E98-CF0F-4BF2-8229-C3FD8DDC794C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CE371E98-CF0F-4BF2-8229-C3FD8DDC794C}.Release|x64.ActiveCfg = Release|Any CPU
+		{CE371E98-CF0F-4BF2-8229-C3FD8DDC794C}.Release|x64.Build.0 = Release|Any CPU
+		{CE371E98-CF0F-4BF2-8229-C3FD8DDC794C}.Release|x86.ActiveCfg = Release|Any CPU
+		{CE371E98-CF0F-4BF2-8229-C3FD8DDC794C}.Release|x86.Build.0 = Release|Any CPU
+		{C947E269-7D43-49ED-99C5-DB0C084B217E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C947E269-7D43-49ED-99C5-DB0C084B217E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C947E269-7D43-49ED-99C5-DB0C084B217E}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C947E269-7D43-49ED-99C5-DB0C084B217E}.Debug|x64.Build.0 = Debug|Any CPU
+		{C947E269-7D43-49ED-99C5-DB0C084B217E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C947E269-7D43-49ED-99C5-DB0C084B217E}.Debug|x86.Build.0 = Debug|Any CPU
+		{C947E269-7D43-49ED-99C5-DB0C084B217E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C947E269-7D43-49ED-99C5-DB0C084B217E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C947E269-7D43-49ED-99C5-DB0C084B217E}.Release|x64.ActiveCfg = Release|Any CPU
+		{C947E269-7D43-49ED-99C5-DB0C084B217E}.Release|x64.Build.0 = Release|Any CPU
+		{C947E269-7D43-49ED-99C5-DB0C084B217E}.Release|x86.ActiveCfg = Release|Any CPU
+		{C947E269-7D43-49ED-99C5-DB0C084B217E}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
